[Refactor] emptyView -> placeholder로 변경
